refactor: updated config installer and templates
