+++ cli/template/config/env-with-next-database.ts
