+++ cli/template/config/env.ts
@@ -1,0 +1,9 @@
+import { createEnv } from '@t3-oss/env-core';
+
+/** ____ STUB: DON'T TOUCH THIS ______ */
+export const env = createEnv({
+  clientPrefix: 'NEXT_',
+  client: {},
+  server: {},
+  runtimeEnv: {},
+});
