+++ cli/template/config/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@acme/config-next-database",
+  "name": "@acme/config",
   "version": "1.0.0",
   "type": "module",
   "license": "MIT",
