+++ cli/template/config/tsconfig.json
