--- cli/template/config/with-next-auth-database/env.ts
@@ -1,31 +1,0 @@
-import { createEnv } from '@t3-oss/env-core';
-import { z } from 'zod';
-
-export const env = createEnv({
-  /*
-   * Specify what prefix the client-side variables must have.
-   * This is enforced both on type-level and at runtime.
-   */
-  clientPrefix: 'NEXT_',
-  client: {},
-  server: {
-    SITE_URL: z.string().url(),
-    GOOGLE_CLIENT_ID: z.string(),
-    GOOGLE_CLIENT_SECRET: z.string(),
-    DATABASE_HOST: z.string(),
-    DATABASE_USERNAME: z.string(),
-    DATABASE_PASSWORD: z.string(),
-    DATABASE_NAME: z.string(),
-    DATABASE_STRING: z.string(),
-  },
-  runtimeEnv: {
-    SITE_URL: process.env.SITE_URL,
-    DATABASE_HOST: process.env.DATABASE_HOST,
-    DATABASE_USERNAME: process.env.DATABASE_USERNAME,
-    DATABASE_PASSWORD: process.env.DATABASE_PASSWORD,
-    DATABASE_NAME: process.env.DATABASE_NAME,
-    DATABASE_STRING: process.env.DATABASE_STRING,
-    GOOGLE_CLIENT_ID: process.env.GOOGLE_CLIENT_ID,
-    GOOGLE_CLIENT_SECRET: process.env.GOOGLE_CLIENT_SECRET,
-  },
-});
