--- cli/template/config/with-next-auth-database/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "@acme/config-next-auth-database",
-  "version": "1.0.0",
-  "type": "module",
-  "license": "MIT",
-  "dependencies": {
-    "@t3-oss/env-core": "^0.3.1",
-    "zod": "^3.21.4"
-  }
-}
