--- cli/template/config/with-next-database/tsconfig.json
@@ -1,4 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "include": ["env.ts"]
-}
